Update `req` method on the REST client
This commit updates the `req` method on the REST client to take an
`Option<&T>`, so callers can use a single method regardless of whether
they need to send a body or not.

The method will probably have to accept headers and query params in the
same way, but so far we don't need them.

Signed-off-by: Radu M <root@radu.sh>