Derive `Clone` for types, remove `mut` signature
This commit adds the `Clone` trait for all current types and removes the
requirement that projects client takes a mutable reference to a project.
It does so by performing an internal mutable clone of the object, whose
metadata can be modified to ensure it is not rejected by the server.

Signed-off-by: Radu M <root@radu.sh>